Refactor directives to provide a more literal SVG template
The directives are now extensions of <svg> and <g> tags. The <svg> is
not created by D3 but the <g> components are instead managed by it with
values set through the directives.

This lets us remove D3Gradient and D3Include as we can just include
<svg> template code directly and let Angular handle most simple
situations.

It is simpler to create the gradient and do includes directly